[refactor] `Triptype.java`를 하위 디렉토리(`type`)로 이동
Motivation:
- 도메인 디렉토리를 정리하기 위해 하위 폴더에 type을 두고 타입 관련 파일들을 넣으려 합니다.

Modification:
- `type` 디렉토리 생성 후 타입 관련 파일 이동
- 파일 이동에 따른 패키지 변경에 대한 변경사항 반영
