+++ src/main/java/com/kdt_y_be_toy_project2/domain/trip/domain/Trip.java
@@ -5,6 +5,8 @@ import java.util.List;
 
 import com.kdt_y_be_toy_project2.domain.itinerary.domain.Itinerary;
 import com.kdt_y_be_toy_project2.domain.model.DateScheduleInfo;
+import com.kdt_y_be_toy_project2.domain.trip.domain.type.TripType;
+import com.kdt_y_be_toy_project2.domain.trip.domain.type.TripTypeConverter;
 
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.AttributeOverrides;
