+++ src/main/java/com/kdt_y_be_toy_project2/domain/trip/dto/TripRequest.java
@@ -2,7 +2,7 @@ package com.kdt_y_be_toy_project2.domain.trip.dto;
 
 import com.kdt_y_be_toy_project2.domain.model.DateScheduleInfo;
 import com.kdt_y_be_toy_project2.domain.trip.domain.Trip;
-import com.kdt_y_be_toy_project2.domain.trip.domain.TripType;
+import com.kdt_y_be_toy_project2.domain.trip.domain.type.TripType;
 import com.kdt_y_be_toy_project2.global.util.DateTimeUtil;
 
 import io.swagger.v3.oas.annotations.media.Schema;
